feat: Implements initial release of errorex
The "errorex" package is a Go package that provides a custom error type with additional information such as error code and details. It also includes a function to register error codes and a function to check if an error is of type "errorex" and if its code matches a particular value.